+++ src/hacks/Global/ShowTaps.cpp
@@ -17,8 +17,8 @@ namespace eclipse::hacks::Global {
         if (touchNodes.empty()) return;
 
         bool fill = config::get<bool>("global.show-taps.fill", true);
-        int radius = (config::get<float>("global.show-taps.scale", 0.5F) * 16.F);
-        int segments = (config::get<float>("global.show-taps.scale", 0.5F) * 32.F);
+        int radius = (config::get<double>("global.show-taps.scale", 0.5F) * 16.F);
+        int segments = (config::get<double>("global.show-taps.scale", 0.5F) * 32.F);
         auto color = config::get<gui::Color>("global.show-taps.color", gui::Colors::WHITE).toCCColor3B();
 
         int stroke = config::get<int>("global.show-taps.stroke", 0);
